Reverted tabbed view and support.v4 stuff
