+++ app/src/main/java/webdevils/webdevilsapp/Employee/EmpApprovedConceptRecyclerViewAdapter.java
@@ -1,9 +1,10 @@
-package webdevils.webdevilsapp.Employee.ManageConcepts.SubmittedTab;
+package webdevils.webdevilsapp.Employee;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
 import android.widget.TextView;
 
 import java.util.LinkedList;
@@ -14,15 +15,17 @@ import webdevils.webdevilsapp.R;
 /**
  * Created by Kevin 04/01/2017
  */
-public class EmpSubmittedConceptRecyclerViewAdapter extends RecyclerView.Adapter<EmpSubmittedConceptRecyclerViewAdapter.ViewHolder> {
+public class EmpApprovedConceptRecyclerViewAdapter extends
+        RecyclerView.Adapter<EmpApprovedConceptRecyclerViewAdapter.ViewHolder> {
 
     private final LinkedList<Concept> mValues;
-    private final EmpSubmittedConceptListFragment.OnSubmittedListFragmentInteractionListener mListener;
+    private final EmpApprovedConceptListFragment.OnApprovedListFragmentInteractionListener mListener;
     public static Concept conceptUnderReview;
     View view;
 
-    public EmpSubmittedConceptRecyclerViewAdapter(LinkedList<Concept> items,
-                                                  EmpSubmittedConceptListFragment.OnSubmittedListFragmentInteractionListener listener) {
+    public EmpApprovedConceptRecyclerViewAdapter(LinkedList<Concept> items,
+                                                 EmpApprovedConceptListFragment.
+                                                         OnApprovedListFragmentInteractionListener listener) {
         mValues = items;
         mListener = listener;
     }
@@ -30,21 +33,26 @@ public class EmpSubmittedConceptRecyclerViewAdapter extends RecyclerView.Adapter
     @Override
     public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
         view = LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.emp_list_submitted_item_card, parent, false);
+                .inflate(R.layout.emp_list_approved_item_card, parent, false);
         return new ViewHolder(view);
     }
 
     @Override
     public void onBindViewHolder(final ViewHolder holder, int position) {
         holder.mItem = mValues.get(position);
-        if(holder.mItem.isApproved()){
-            //change R.layout.emp_list_approved_item_card here ???
-        }
+
         conceptUnderReview = holder.mItem;
-        holder.mCreatedByView.setText("Concpet Created By: " + mValues.get(position).getUserThatCreatedThisConcept().getUserName());
+        holder.mCreatedByView.setText("Concpet Created By: " + mValues.get(position).
+                getUserThatCreatedThisConcept().getUserName());
         holder.mTitleView.setText(" Title: " + mValues.get(position).getTitle());
         holder.mEmpReviewedStatus.setText("Status: " + mValues.get(position).getStatus());
         holder.mConceptType.setText("Concept Type: " + mValues.get(position).getType());
+
+        if(mValues.get(position).isSticky()){
+            holder.mImageView.setImageResource(R.mipmap.note_pinned);
+        }else{
+            holder.mImageView.setImageResource(R.mipmap.blank_note_pinned);
+        }
 
         holder.mView.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -52,7 +60,7 @@ public class EmpSubmittedConceptRecyclerViewAdapter extends RecyclerView.Adapter
                 if (null != mListener) {
                     // Notify the active callbacks interface (the activity, if the
                     // fragment is attached to one) that an item has been selected.
-                    mListener.onSubmittedConceptListFragmentInteraction(holder.mItem);
+                    mListener.onApprovedConceptListFragmentInteraction(holder.mItem);
                 }
             }
         });
@@ -85,6 +93,7 @@ public class EmpSubmittedConceptRecyclerViewAdapter extends RecyclerView.Adapter
         public final TextView mTitleView;
         public final TextView mEmpReviewedStatus;
         public final TextView mConceptType;
+        public final ImageView mImageView;
         public Concept mItem;
 
         public ViewHolder(View view) {
@@ -95,6 +104,7 @@ public class EmpSubmittedConceptRecyclerViewAdapter extends RecyclerView.Adapter
             mTitleView = (TextView) view.findViewById(R.id.concept_title);
             mEmpReviewedStatus = (TextView) view.findViewById(R.id.concept_employee_reviewed_status);
             mConceptType = (TextView) view.findViewById(R.id.concept_type);
+            mImageView = (ImageView) view.findViewById(R.id.sticky);
 
         }
