+++ app/src/main/java/webdevils/webdevilsapp/Employee/EmpConceptReviewFragment.java
@@ -3,13 +3,12 @@ package webdevils.webdevilsapp.Employee;
 import android.app.Activity;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import android.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import common.Concept;
-import webdevils.webdevilsapp.Employee.ManageConcepts.SubmittedTab.EmpSubmittedConceptRecyclerViewAdapter;
 import webdevils.webdevilsapp.R;
 
 /**
