--- app/src/main/java/webdevils/webdevilsapp/Employee/ManageConcepts/ApprovedTab/EmpApprovedConceptListFragment.java
@@ -1,105 +1,0 @@
-package webdevils.webdevilsapp.Employee.ManageConcepts.ApprovedTab;
-
-import android.app.Activity;
-import android.content.Context;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import common.Concept;
-import webdevils.webdevilsapp.Employee.ListContent.ConceptListContent;
-import webdevils.webdevilsapp.R;
-
-/**
- * A fragment representing a list of Items.
- * <p/>
- * Activities containing this fragment MUST implement the {@link OnApprovedListFragmentInteractionListener}
- * interface.
- */
-public class EmpApprovedConceptListFragment extends Fragment {
-
-    private static final String ARG_COLUMN_COUNT = "column-count";
-    private int mColumnCount = 1;
-    private OnApprovedListFragmentInteractionListener mListener;
-    public static Concept conceptUnderReview;
-
-    /**
-     * Mandatory
-     */
-    public EmpApprovedConceptListFragment() {
-    }
-
-    public static EmpApprovedConceptListFragment newInstance(int columnCount) {
-        EmpApprovedConceptListFragment fragment = new EmpApprovedConceptListFragment();
-        Bundle args = new Bundle();
-        args.putInt(ARG_COLUMN_COUNT, columnCount);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        if (getArguments() != null) {
-            mColumnCount = getArguments().getInt(ARG_COLUMN_COUNT);
-        }
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.emp_list_approved, container, false);
-
-        // Set the adapter
-        if (view instanceof RecyclerView) {
-            Context context = view.getContext();
-            RecyclerView recyclerView = (RecyclerView) view;
-            if (mColumnCount <= 1) {
-                recyclerView.setLayoutManager(new LinearLayoutManager(context));
-            } else {
-                recyclerView.setLayoutManager(new GridLayoutManager(context, mColumnCount));
-            }
-            recyclerView.setAdapter(new EmpApprovedConceptRecyclerViewAdapter(
-                    ConceptListContent.APPROVED_ITEMS, mListener));
-        }
-        return view;
-    }
-
-
-    @Override
-    public void onAttach(Activity activity) {
-        super.onAttach(activity);
-        if (activity instanceof OnApprovedListFragmentInteractionListener) {
-            mListener = (OnApprovedListFragmentInteractionListener) activity;
-        } else {
-            throw new RuntimeException(activity.toString()
-                    + " must implement OnRejectedListFragmentInteractionListener");
-        }
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        mListener = null;
-    }
-
-    /**
-     * This interface must be implemented by activities that contain this
-     * fragment to allow an interaction in this fragment to be communicated
-     * to the activity and potentially other fragments contained in that
-     * activity.
-     * <p/>
-     * See the Android Training lesson <a href=
-     * "http://developer.android.com/training/basics/fragments/communicating.html"
-     * >Communicating with Other Fragments</a> for more information.
-     */
-    public interface OnApprovedListFragmentInteractionListener {
-        void onApprovedConceptListFragmentInteraction(Concept concept);
-    }
-}
