--- app/src/main/res/layout/emp_list_rejected.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.RecyclerView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/unreviewed_list"
-    android:name="webdevils.webdevilsapp.Employee.EmpConceptReviewFragment"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/list_selector"
-    tools:context="webdevils.webdevilsapp.Employee.ManageConcepts.SubmittedTab.EmpSubmittedConceptListFragment"
-    tools:listitem="@layout/emp_list_rejected_item_card"/>
