--- app/src/main/res/layout/emp_list_rejected_item_card.xml
@@ -1,60 +1,0 @@
-<android.support.v7.widget.CardView
-    xmlns:card_view="http://schemas.android.com/apk/res-auto"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/card_view"
-    android:layout_width="match_parent"
-    android:layout_height="80dp"
-    card_view:cardCornerRadius="4dp"
-    android:layout_margin="5dp"
-    card_view:cardUseCompatPadding="true">
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="80dp">
-
-                <TextView
-                    android:id="@+id/concept_title"
-                    android:layout_width="match_parent"
-                    android:layout_height="20dp"
-                    android:background="#960C11"
-                    android:text="Title"
-                    android:gravity="center_vertical"
-                    android:textColor="@android:color/white"
-                    android:textSize="14dp"/>
-
-                <TextView
-                    android:id="@+id/concept_created_by"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_below="@id/concept_title"
-                    android:layout_marginLeft="5dp"
-                    android:layout_marginTop="10dp"
-                    android:elevation="7dp"
-                    android:gravity="center_vertical"
-                    android:text="Memeber Name"
-                    android:textSize="10dp" />
-
-                <TextView
-                    android:id="@+id/concept_employee_reviewed_status"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="Status"
-                    android:gravity="center_vertical"
-                    android:textSize="10dp"
-                    android:layout_below="@id/concept_created_by"
-                    android:layout_marginTop="10dp"
-                    android:layout_marginLeft="5dp"/>
-
-                <TextView
-                    android:id="@+id/concept_type"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="Concept Type"
-                    android:textSize="10dp"
-                    android:layout_marginRight="200dp"
-                    android:layout_alignBaseline="@+id/concept_created_by"
-                    android:layout_alignBottom="@+id/concept_created_by"
-                    android:layout_centerHorizontal="true" />
-
-        </RelativeLayout>
-</android.support.v7.widget.CardView>
