+++ app/src/main/res/layout/emp_list_submitted.xml
@@ -7,5 +7,5 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/list_selector"
-    tools:context="webdevils.webdevilsapp.Employee.ManageConcepts.SubmittedTab.EmpSubmittedConceptListFragment"
+    tools:context="webdevils.webdevilsapp.Employee.EmpSubmittedConceptListFragment"
     tools:listitem="@layout/emp_list_submitted_item_card"/>
