+++ app/src/main/res/menu/emp_navigation_menu.xml
@@ -10,7 +10,7 @@
         <item
             android:id="@+id/manage_concepts"
             android:icon="@drawable/ic_menu_manage"
-            android:title="Manage Concepts" />
+            android:title="Make Concepts Sticky" />
         <item
             android:id="@+id/nav_update_member_news"
             android:icon="@drawable/ic_menu_gallery"
